+++ R/alma-search.R
@@ -21,9 +21,7 @@
 #' on_monday <- weekly() %>% recur_on_day_of_week("Monday")
 #'
 #' # On the 12th of the month, or on Mondays
-#' rb <- runion() %>%
-#'   add_rschedule(on_12th) %>%
-#'   add_rschedule(on_monday)
+#' rb <- runion(on_12th, on_monday)
 #'
 #' alma_search("2019-01-01", "2019-01-31", rb)
 alma_search <- function(from, to, rschedule, inclusive = TRUE) {